Formatted the price with tax result to a better and readable way

+++ item_model.py
@@ -7,4 +7,4 @@ class Item(BaseModel):
     name: str
     description: Optional[str] = None
     price: float
-    tax: float
+    tax: int
